Fix skipping upstream syncs with no changes when landing (#459)
* Don't try to rebase upstream PRs before merging.

This was used to try to improve the error messages, but is unnecessary and
confusing since it changes the local SHA1 to differ from the head of the PR
branch.

* Check the reflog when looking for upstream syncs that don't need to be relanded

This allows us to work in the case that the PR was rebased locally but the rebase not
pushed, which was a thing we were doing. There is an edge case where this will fail if we
remove a commit from the PR, but it's also likely a human would merge that case incorrectly,
and it doesn't happen often so it seems like a lesser evil.