Remove .mozbuild wpt cache
This prevents the cache becoming a space leak, and avoids a somewhat
common error we've seen with the cache becoming corrupted. The reasons
for this are unclear, but the performance implications of simply not
using the cache are well within the bounds of what's reasonable to get
a simple fix.