Add results notifications based on wpt.fyi data
wpt.fyi has results for multiple browsers so it gives some insight
into how Firefox does on an import compared to Chrome and Safari. This
can be useful for priotitising tests that should be investigated
quickly.

Initially the approach is just to create a results string from wpt.fyi
data and a second string from the gecko data and concatenate
them. This is probably a little over-verbose as we might hope there's
some overlap in the results. However we can improve the display once
the idea is proven out.