Only perform a single try push per downstream PR
In order to conserve resources, and simplify the handling of downstream syncs, only perform a single try push, after the upstream PR has merged. For PRs that involve tests this will be a stability try push with 5 repetitions of the affected tests, for infrastructure changes this will be a single iteration of the infra tests.

This change does make a tradeoff in reducing the best-case latency for landing since PRs now have to do their try run after they have merged never before. But we wasted lots of resources doing try runs for things that underwent significant change after approval.