Change the sequencing of marking landing try push as complete
Previously we only marked the try push as complete once the results
had been used and subsequent actions like rebase had been
completed. But that meant that a failure in those subsequent actions
caused the metadata to get re-updated.

Instead mark the try push as complete right after completing the
metadata updates, and ensure that subsequent runs of the code don't
repeat the metadata update for a complete try push.