Enable pushing to phabriactor using moz-phab
Now that inbound is gone, we have to push to autoland. But pushing to
autoland is hard because we struggle to find a time when it's open and
we don't hit a conflict with lando. Instead what we'd like is to end
up in the lando queue, but so far lando only knows how to read from
phabricator. So as a temporary solution we adopt the following
process:

* Push the landing to Phabricator using moz-phab
* Needinfo the wpt sync user to manually push the commits to lando and
  invoke lando (ignoring all the warnings about unlanded code).