Add support for auto-updating bug metadata
If we link a test result to a bug and then changes are made to the
bug, we may need to update the metadata. This implements a background
task to examine linked bugs, check for relevant changes, and make
updates if required. The two kinds of update initially supported are:
 * If a bug is resolved as duplicate, we update the link to point to
 the bug it's a duplicate of
* If a bug is resolved as invalid we remove the link on the basis that
  an invalid bug can't explain a test result.

In the latter case we also comment on the bug to explain what happened.