Ensure that all lookups in ProcessNameIndex are done with strings
There's a general confusion in the codebase about whether
obj_ids (i.e. PR numbers and bug numbers) are represented by integers
or strings. In ProcessName objects they are always strings, but
sometimes in other contexts they aren't. This warrants a general
solution, but there was a specific problem with lookups in the
ProcessNameIndex passing in an integer obj_id and not finding
anything, so force the type to be str inside the get() function.