Set gecko_base when doing gecko_rebase
The gecko_rebase() function previously forgot to set the
gecko-base property of the sync, so the commit range would
end up being too large. This only mattered in practice when there
were other wpt-sync commits on autoland, but not on central.

Change things so that when we do the rebase we always set the new base.
There is still a failure mode where the rebase fails and we don't
manually redo the rebase, so the stored base would end up as
autoland rather than central, but the actual base would be central.
Even in this pathological case it's pretty unlikely that anything
actually bad will happen; the worst cas scenario is that there are
commits on central that haven't been applied to autoland that look
like part of the sync.