Create PRs when pushing to wpt-metadata repo
We were previously doing metadata updates by pushing directly to
master in (a fork of) the wpt-metadata repository. However because the
repo has CI checks that must pass before merging PRs to ensure that
the data is valid we actually want to create a PR first. This adds an
(optional) PR creation step when writing metadata.