Check for taskgroup completion when tasks finish
There is a failure mode we've encountered where we get some builds failing,
and the resulting taskgroup doesn't complete until all the unscheduled tasks
time out a day later. But we want to handle that case upfront, as soon as we
can know that the remaining tasks aren't going to run.

To do this, we process all tasks rather than just decision tasks. When a
non-decision task completes, we load all the tasks from the taskgroup and
check for completeness ignoring anything that's unscheduled. This has some
performance cost, so we are careful to only look at tasks we scheduled
ourselves.