Use push to central for base of logs, not merge-base (#663)
The code to get the baseline logs for central assumed that it
could get the central push using the merge-base of the commit with
central. But this doesn't work in the case where the sync was
rebased onto autoland and autoland merged into central. In this
case the merge-base will be the tip of autoland at the point where
we did the rebase, and that isn't usually a push head from central.

Instead we need to take the merge-base and ask hg what the central
push containing that commit was. Happily that's almost exactly the
same as code at [1] so the implementation can be mostly copied.

[1] https://searchfox.org/mozilla-central/source/testing/web-platform/manifestdownload.py#74

Co-authored-by: Karl Dubost <karl+github@la-grange.net>