Make it possible to build image with Python 3
This adds a --py3 flag to the `run_docker_dev.sh` script that means
the image is built with Python 3 as the wptsync Python. This can then
be used to test and run wptsync under Python 3. The Python 3 image is
given a different docker tag to dintinguish it from the Py 2 image.

A further flag --test was added to `run_docker_dev.sh` which skips the
ssh key setup steps. This makes the script suitable for local testing
or testing on travis where ssh keys aren't available (or required).

Note that because the shell script is terrible, the flags must be
passed in order.

As part of the Dockerfile changes, a significant refactor was
performed. In particular the wpt-sync code is no longer copied into
the image, but must be mounted on the host. This was the configuration
we were using in practice anyway, and making this change improves the
docker build speed considerably.

As a result of this, various files required to set up the rest of the
image have to be explictly copied into the image at build
time (notably requirements files).

In addition, the docker image was changed to remove the
virtualenv. Since we have control of the full python setup inside the
image, also adding a virtualenv was unnecessary additional complexity.

Some additional simplification of the dockerfile came along for the ride.