Switch to using check-run events for determining when PRs are mergable
Upstream wpt has switched the taskcluster integration to use checks
rather than the older status API. This means that we are no longer
able to detect when a PR is mergeable by checking the combined status
of the PR; instead we need to look at the check runs and schedule the
events on the basis of check runs.

pygithub (at least the py2 compatible version) doesn't support the
check_runs API so we have to implement the relevant classes ourselves
extending pygithub. This turns out to be a little easier than just
using the http api directly particuarly because we get pagination
support for free.

To decide if a sync is likely mergable, we check if all the required
status checks have completed successfully. The listener is changed to
listen to check-run events where the run has completed.