Some initial type fixes in commit.py
This is rather critical since much of the rest of the code interacts
with git objects via this module. For data retrieved from git which
we don't fully control, it's important to work with bytes everywhere
since patches, commit messages, and metadata can all contain non-utf-8
sequences and we've previously had bugs related to this.

For tags we are currently taking a risk and assuming Text, because
the number of tags is generally small and the chance of someone
creating one that isn't valid utf-8 is also small.