Ensure mach commands use correct, independent state
Running `mach` now requires first running `mach
create-mach-environment`, which creates a virtualenv. We need to do
that for each worktree and we don't want to share state between
worktrees, since recreating the mach environment while a command is
running may cause problems. So instead we add lifecycle events to
worktree creation and deletion, and use those events to first create a
state directory specific to the worktree and then later delete it. To
ensure the correct state directory is used we also need to set the
MOZBUILD_STATE_PATH environment variable each time we invoke mach.