Ensure mach environment dir exists for each mach command
Creating once per worktree is efficient but doesn't work in some
cases, notably when a branch is rebased from a revision where no
environment is required to one where it is. The simplest solution to
this seems to be to always check that it exists, but it does mean
hardcoding in some assumptions.