Set rabbitmq consumer timeout to 10 hours
A recent rabbitmq update broke the sync because it set the task
timeout to 15 minutes, after which the node would disconnect and all
task processing would stop.

We can in some cases have quite long-running tasks. In particular the
way retrigger works currently is as a single task (this should perhaps
change), and that can take longer than an hour. So set the timeout to
10 hours to be reasonably sure that the node isn't going to get
shutdown halfway through a task.