Don't add a cinnabar property to Repo objects
We were previously adding a custom `cinnabar` property to repo objects
representing gecko repositories. mypy wasn't happy with this, so
instead switch to caching the cinnabar object in a dict and looking it
up when required.

Note: this could leak if we discard Repo objects. But we already have
the same problem with the pygit2 objects, so fixing that can be done
in the future.