Handle the possibility of a missing reflog
When landing PRs that came from mozilla-central we check if the are
upstream commits are identical to the ones we already pushed, because
we know there's no point in re-applying those.

One of the codepaths uses the reflog to check if the upstream commits
are the same as any previous head of the PR branch. But it's not
guaranteed that a reflog exists for a branch. In particular for very
old branches it may have been gc'd. gitpython doesn't handle that
situation very well, and raises an exception which suggests a bug.

We can handle that exception and just fallback to the code which
assumes that the commits may have changed (even though this is likely
to generate a merge conflict that has to be handled manually).