Handle _lock on Mock objects
unittest.Mock got an _lock attribute to synchronise access to the Mock
in multithreaded cases. Unfortunately this clashes with our use of
_lock on syncs to ensure we don't allow more than one process to
mutate a sync at once.

For the cases where we're mocking whole Sync objects in tests, just
overwrite the _lock class attribute with another mock; this seems to
be enough for the tests to pass again.