Handle labels in META.yml files
Previously we only correctly handled metadata entries which had a
"url" key, which generally speaking labels do not. In practice that
meant we failed to notify for test changes in directories with labels.

This change adds a specific type to represent labels and ensure that
they're correctly roundtripped through the update process.

This matches the data representation used in other tooling, but we
will still need to update the code if a new kind of metadata is added.