Enable ordering of ProcessName class
In worktree.py we have a list of tuples of the form (mtime,
ProcessName, path) which we try to sort. Usually this works fine, but
in the case where two worktrees have the exact same mtime it breaks
because ProcessName isn't sortable. To solve this, make the
ProcessName class implement __lt__.